// apps/api/src/common/enums.ts
export enum UserRole {
  PUBLIC = 'PUBLIC',
  USER = 'USER',
  STUDENT = 'STUDENT',
  SCHOOL_ADMIN = 'SCHOOL_ADMIN',
  SUPERADMIN = 'SUPERADMIN',
}

export enum EmploymentType {
  NONE = 'NONE',
  STUDENT = 'STUDENT',
  APPRENTICE = 'APPRENTICE',
  INTERN = 'INTERN',
}

export enum OnboardingStep {
  STEP0 = 'STEP0',
  STEP1 = 'STEP1',
  STEP2 = 'STEP2',
  DONE = 'DONE',
}

export enum ContactType {
  EMAIL = 'email',
  LINKEDIN = 'linkedin',
  URL = 'url',
  OTHER = 'other',
}